Refactor the project, again
This commit attempts to improve the S3Proxy package.
The style was chosen after looking at other open source go projects like
Hugo and Consul.
