+++ .gitignore
@@ -1,1 +1,1 @@
-S3Proxy/S3proxy
+bin/
