+++ S3Proxy/configure.go
@@ -1,0 +1,7 @@
+package S3Proxy
+
+import "os"
+
+func Configure() {
+   os.Mkdir(Options.CacheDir, 0700)
+}
