+++ S3Proxy/handlers.go
