+++ S3Proxy/logging.go
